Update vim-lsc API, mock source from Yggdrasil
Update the vim-lsc API, after the lsc#server#call() function was
removed.

When making an LSP request from within an Yggdrasil buffer, mock a
source file. This because some LSP clients accept requests only from
what seems to be a legit source file.

Replace the previous, brittle solution of jumping to the previous window,
that may not work if the user moved between different windows before
expanding a node in the Yggdrasil buffer.